refactor(game): move resolution to SolidTile
