+++ src/game/engine/index.js
@@ -2,4 +2,3 @@ export * from './BoxCollider';
 export * from './GameEngine';
 export * from './GameObject';
 export * from './GameScene';
-export * from './physicsUtils';
