+++ src/game/objects/SolidTile.js
@@ -22,6 +22,28 @@ class SolidTile extends GameObject {
   onUpdate(frame) {
     frame.buffer.draw(this.position, this.#sprite);
   }
+
+  /**
+   * @param {GameObject} object
+   * @param {import('../shared').Collision} collision
+   * @param {number} restitution https://en.wikipedia.org/wiki/Coefficient_of_restitution
+   * @return {import('../shared').Vector} normal vector from the solid to the object
+   */
+  static resolveCollisionWith(object, collision, restitution) {
+    // eslint-disable-next-line no-param-reassign
+    object.transform.position = object.transform.position.add(collision.resolutionVector);
+
+    const normal = collision.resolutionVector.normalized();
+    if (object.rigidbody.velocity.dot(normal) <= 0) {
+      // Source: http://www.chrishecker.com/images/e/e7/Gdmphys3.pdf
+      const j = -(1 + restitution) * object.rigidbody.velocity.dot(normal);
+      const impulse = normal.scale(j);
+
+      object.rigidbody.addVelocity(impulse);
+    }
+
+    return normal;
+  }
 }
 
 export { SolidTile };
